pyproject.toml
[tool.poetry]
name = "api-testing-cotodigital"
version = "0.1.0"
description = "API testing using Python, Requests and Pytest for CotoDigital API"
authors = ["Gomez Sergio Alberto <sergio.alberto.gomez@coto.com.ar>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
pytest = "^7.4.2"
assertpy = "^1.1"
cerberus = "^1.3.5"
pytest-html = "^4.0.2"
selenium = "^4.14.0"
pre-commit = "^3.5.0"
allure-pytest = "^2.13.2"
sqlalchemy = "^2.0.23"
oracledb = "^1.4.2"
pymssql = "^2.2.10"
python-dotenv = "^1.0.0"
invoke = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
